+++ packages/cli/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
 - CLI can now be imported from another module.
+- `generate` cleans the output dir before compiling sources.
 
 ## 0.0.1 - 2024-04-03